fix(Products): Corrige retorno de produtos e adiciona retorno de categorias em products

+++ src/cache/company/utils/invalidateCache.ts
@@ -6,9 +6,13 @@ export const invalidateCache = async (companyId: string, module: string) => {
     const stream = client.scanIterator({ MATCH: patternWithFilters }) //varredura paginada de chaves, retorna aos poucos, não tudo de uma vez.
     let count = 0
 
-    for await (const key of stream) {
-        if (!key) continue
-        await client.del(key)
-        count++
+    for await (const keyOrKeys of stream) {
+        const keys = Array.isArray(keyOrKeys) ? keyOrKeys : [keyOrKeys]
+        for (const key of keys) {
+            if (!key || key.length === 0) continue
+            await client.del(key)
+            count++
+        }
     }
+
 }
